Modified draftscraping file
Scrapes basketball.realgm.com versus wikipedia. Prompts user for date
range as well as the name of the csv file we're going to write to. This
gives the  names of undrafted players during each year so produces a
much larger dataset. Need to implement error checking but runs
accurately and is very quick.