Gathers the urls and writes stats for each player
Introduced a menu function to prompt user for csv file we'll write our
stats out to, then gather the specified url for each player and store
all the urls in a text file. Then it also scrapes the urls to get the
stats and append them to the csv file.