Initial Clustering and Graph of NBA Draft Data
Beginning to cluster in different groups with the data. Currently it clusters everything but the name, team, prev team, position, etc and each time it clusters players in groups with the same number of players even if the clusters are numbered differently. Prints out one graph at the end. Will need to update and choose a different type of graph so we can see the numbers properly.